+++ plugins/action/wireless_sensor_test_results_info.py
@@ -26,8 +26,8 @@ argument_spec = dnac_argument_spec()
 # Add arguments specific for this module
 argument_spec.update(dict(
     siteId=dict(type="str"),
-    startTime=dict(type="int"),
-    endTime=dict(type="int"),
+    startTime=dict(type="float"),
+    endTime=dict(type="float"),
     testFailureBy=dict(type="str"),
     headers=dict(type="dict"),
 ))
